Database connection type reworked
The Hibernate initialization has been reworked for catalog and request
DBs as it was not generic enough for those who want to use it as a
maven library.
The parameter -Dmso.db that must be provided to the JVM startup can now
be set to anything but the code will fail to load the DB if the filename
do not exist in the WAR resources.
There was no way to extend the RequestDatabase class as it was final

Change-Id: Ib9496ed027ebcf58edcd853e263d82d2c00b8641
Signed-off-by: Determe, Sebastien (sd378r) <sd378r@intl.att.com>