Update patch set 1
Patch Set 1:

Hmm, ok.  In OPEN-O and OpenDaylight we don't have developers build against staging artifacts; it is assumed that any artifacts with release version labels are final, immutable, and already available from the release repos.   This prevents the issue of an artifact with the same release version label but having different content on different environments due to timing differences in when it was downloaded.

We'll roll with the "build against staging" approach for now, but we may want to reconsider this approach later.

Patch-set: 1
Label: Code-Review=0