Failed to deploy so artifacts in nexus (again)
There are Linux Foundation rules (enforced by nexus) that I still
don't really understand.

I'm convinced this is absolutely true:

The groupId of every artifact in a project MUST begin with:
	org.openecomp.<project>
	   --or--
	org.onap.<project>

The top-level artifact is no exception.  So for example, this is
NOT allowed:

	<groupId>org.openecomp</groupId>
	<artifactId>so</groupId>

I'm not so sure about this:

The artifactId in the top level pom must match the project name,
so for "so", the artifact MUST be:

	<groupId>org.openecomp.so</groupId>
	<artifactId>so</groupId>

Consequently, the path to the top level "so" artifact will be:

	org/openecomp/so/so/1.1.0-SNAPSHOT/so-1.1.0-SNAPSHOT.pom

It seems rather unfortunate to have "/so/so/" in the path.  Perhaps
the "so" project repositories could be restructured to be more like
other projects', but this time, it's impractical.

Issue: SO-21
Change-Id: Id74cdc4fd47bda1d1ec5aff1d01726891f544e06
Signed-off-by: Rob Daugherty <rd472p@att.com>