Failed to deploy so artifacts in nexus (again)
There are Linux Foundation rules (enforced by nexus) that I still
don't really understand. I'm convinced this is absolutely true:

The groupId of every artifact in a project MUST begin with:
	org.openecomp.<project>
	   --or--
	org.onap.<project>

The top-level artifact is no exception.  So for example, this is
NOT allowed:

	<groupId>org.openecomp</groupId>
	<artifactId>so</groupId>

Here's what I'm trying now.  The top level pom will contain:

        <groupId>org.openecomp.so</groupId>
        <artifactId>so-parent</artifactId>

Child modules will contain:

        <parent>
                <groupId>org.openecomp.so</groupId>
                <artifactId>so-parent</artifactId>
                <version>1.1.0-SNAPSHOT</version>
        </parent>

        <groupId>org.openecomp.so</groupId>
        <artifactId>some-child-artifact</artifactId>

Note that the groupId for the direct child module will be the
same as the groupId for its parent.

Issue: SO-21
Change-Id: I0d3cd2eb7a1883e23e3c0878ee7fa3dd4a7d55b2
Signed-off-by: Rob Daugherty <rd472p@att.com>