Fix to pass transaction id to AAI
When SO is calling AAI REST API in groovy code, it is always passing new
X-TransactionID. As per EELF logging guidelines, this id is used for
tracking transaction across components. This fix checks if there is
already an existing transaction id in MDC, if yes then pass same. If not
found, then create one and set in MDC for use in next API calls in same
transaction.

Change-Id: If6e33b77d9615f4887e1c1fc367de4268eecc340
Issue-ID: SO-78
Signed-off-by: anupmarathe <anup.marathe@amdocs.com>