SO Failed Robot Health Check
There were several issues.  The wrong jackson provider was being selected,
which caused the exception, so we fixed that. Then we discovered that a
a URN mapping was missing, so we added that.  Then we found that the
healthcheck wouldn't work because it's trying to check the default
process engine, which does not exist.  We commented out that part of
the healthcheck to be able to resolve this high priority issue.  A
new low priority issue will need to be created to follow up.

Change-Id: I327612a147333eeecd44915942cecd13959c4717
Issue-id: SO-139
Signed-off-by: Rob Daugherty <rd472p@att.com>