Changed copyright information
Changed copyright info in the listed files.
For csar files, files were unzipped using unzip csarFileName
The text files were modified:
The license info was added in the beginning of the file, not at the end.
And then zipped again: zip csarFileName -r *
For Xsd files, the information was added within commment section

Change-Id: Ifc7c9594f948f98077c4364f4b9631fcf6cb1c32
Issue-Id: SO-252
Signed-off-by: Shailendra Borale <sb8915@att.com>