Changed aaiEndpoint to point to new server name
aaiEndpoint was pointing to the vm1 server which is aliased to the
new server.  However, as the new server's certificate does not
include the alias as one of its alternate names, things were unable
to connect.  The endpoint has now been changed to match the new
server name, thus the SSL ceritifcate should match now.  Note: for
this to work, the new server name must be found in DNS (or /etc/hosts).

Change-Id: Ic865b8dbdf2d182a39529dd56625e4e9a9d3fce5
Issue-Id: SO-249
Signed-off-by: Jim Hahn <jrh3@att.com>