Update patch set 2
Patch Set 2:

Build Failed 

https://jenkins.onap.org/job/so-master-release-version-profile-java-daily/67/ : FAILURE

Build failed to find artifact org.openecomp.so:cockpit-urnmap-plugin:jar:1.1.0 needed for sub-project org.openecomp.so:MSOCockpit:pom:1.1.0. Check correct groupId, artifactId and version. This may also indicate a problem with nexus. Another possibility is that this is a -SNAPSHOT version and it has expired from nexus -- in that case check with the project which provides the artifact. ( https://jenkins.onap.org/job/so-master-release-version-profile-java-daily/67/ )

Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/so-master-release-version-profile-java-daily/67

https://jenkins.onap.org/job/so-master-docker-version-java-daily/156/ : FAILURE

Build failed to find artifact org.openecomp.so:cockpit-urnmap-plugin:jar:1.1.0 needed for sub-project org.openecomp.so:MSOCockpit:pom:1.1.0. Check correct groupId, artifactId and version. This may also indicate a problem with nexus. Another possibility is that this is a -SNAPSHOT version and it has expired from nexus -- in that case check with the project which provides the artifact. ( https://jenkins.onap.org/job/so-master-docker-version-java-daily/156/ )

Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/so-master-docker-version-java-daily/156

Patch-set: 2