Clean up Process Engine selection logic
Several failed attempts to split the BPMN application into
multiple applications with separate camunda process engines
have left a mess of confusing classes and process engine
definitions.

In the Amsterdam release, there should be only one BPMN
application war.  This is MSOInfrastructureBPMN.
MSOCommonBPMN should not be deployed as a separate
application.  Its classes are compiled into a jar and
this is included inside MSOInfrastructureBPMN.

The MSOInfrastructureBPMN application should use the
"default" process engine.

WorkflowAsyncInfrastructureResource and
MSOCommonApplication classes are not needed.

Issue: SO-243
Change-Id: Ifdb3b33541346b561a16361d1aa791e8342a34fa
Signed-off-by: Rob Daugherty <rd472p@att.com>