Update patch set 2
Patch Set 2:

I question the semantics here.  Having all the methods synchronized will result in horrific performance.  Would it be better just to use a ConcurrentHashmap and be unsynchronized?  Also, if clients of CloudConfig are not expected to change the returned map, it should be returned using Collections.unmodifiableMap.

Patch-set: 2
Label: Verified=0