Update patch set 1
Patch Set 1:

Hi Rob,
I think (1) is pretty much agreed upon:
See https://wiki.onap.org/display/DW/Java+code+style - recommended for Amsterdam by the TSC.

I agree with (2) and (3) and I think the sooner we start doing so the better.

As an FYI surefire junit and gerrit settings warn, red highlight on use of tabs and surefire junit on bad indentation. If we revert this I'd ask that we change surefire settings for SO to ignore those issues. I believe the surefire settings are currently inherited from o parent and meant, it seems, to be ONAP wide.

Code Style & Standards have not been enforced in some projects, thus the large change here.

Just an FYI of why they change was pushed. The logs I downloaded to debug verify failures had 30k plus lines of warnings:
"The build logs are flooded with warnings due to incorrectly formatted test code. In a recent jenkins log there were 25,577 matches to "warning: Line contains a tab character.", and 9,766 matches for "warning: 'method def' child have incorrect indentation". This patch should reduce logs by my than ~34,000 lines on each build.

Examples of warnings that this patch fixes:
/w/workspace/so-master-merge-profile-java/bpmn/MSOInfrastructureBPMN/src/test/java/org/openecomp/mso/bpmn/infrastructure/CreateVfModuleVolumeInfraV1Test.java:177:1: warning: Line contains a tab character.
/w/workspace/so-master-merge-profile-java/bpmn/MSOInfrastructureBPMN/src/test/java/org/openecomp/mso/bpmn/infrastructure/CreateVfModuleVolumeInfraV1Test.java:178: warning: 'method def' child have incorrect indentation level 16, expected level should be 8.

Search an example log in Jenkins for "warning: Line contains a tab character" here: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/so-master-verify-profile-java/1024/console.log.gz"

I'm sympathetic that you have a large chunk of code that conflicts with this and thus you want to revert it to make it easier for you to merge your code. That sounds like the wrong way to deal with an issue like this but we all have our deadlines.

I'll just say that this was a couple of hours to try to make the code base a little friendly for the average person working with it, and with the build logs it creates.

Patch-set: 1