Update patch set 1
Patch Set 1:

> Please fix this change (rebase if needed) and submit it again, or
 > abandon it.

Thanks Rob. I'm working on it, hopefully pushing today. I wanted to make it public as to the path I was taking. Since it's a draft it should not affect the code review/merge deadline.

I'll add reviewers when I want the change reviewed (it will pass verify at that time).

Patch-set: 1
Label: Code-Review=0