Update patch set 8
Patch Set 8:

Apparently snapshot builds are not enable for SO repositories. We need a new release version, but it also seems prudent to enable snapshot creation and storage in nexus as a temporary workaround and future convenience. Working to enable snapshots for so/libs here: https://gerrit.onap.org/r/#/c/38693/

Patch-set: 8