Update patch set 8
Patch Set 8:

(1 comment)

This is blocked until either there is a new release version of so/libs or snapshots are enabled for so/libs (https://gerrit.onap.org/r/#/c/38693). The former would be best, but the later could be a short term fix and it seems like snapshots should be enabled anyway. Then we could change the the dependency on so/libs and fix the build.

Patch-set: 8
Reviewer: Gerrit User 667 <667@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0