Update patch set 3
Patch Set 3:

> > > Curent implementation invokes overridden methods in constructor
 > > > DmaapClient class, what is a bad practice. I moved this methods
 > > to
 > > > initProperties method. Builder is made to create object, and
 > then
 > > > invoke initProperties method.
 > >
 > > This is an abstract class and this issue does not apply to this
 > > case. I agree it is strange but it does not warrant a solution
 > such
 > > as this.
 > 
 > It doesn't matter if class is abstract or not. Constructor should
 > never call non-static methods, especially abstract ones (and even
 > calling static ones is bad), because the object is not fully
 > initialized. This can easly lead to dangerous and unpredictable
 > situations, like accessing member that's initialized in
 > non-abstract class.

Please show me how to instantiate and instance of an abstract class and we can continue with the point you're trying to make. I agree with you that it's not the most accepted way to solve this particular problem, but we have little risk here. If you want to talk about risk, it seems mighty easy with your implementation for the caller to forget to initialize the required properties for the DmaapConsumer and DmaapPublisher. If only there was a way to guarantee that our objects completed all their required initialization before being used....wait a minute.

Patch-set: 3