Update patch set 3
Patch Set 3:

> Programming is about trade offs. I find developers are more likely
 > to forget to call some "init" method themselves than incorrectly
 > create an implementation of an abstract class.

Ok, to be clear, I know the init method is not the best option (however is far away better than calling abstract methods in constructor), this the first step for better refactoring. 
Anyway, I set constructors private, and init method is protected and invoked from build method of builder. Look more carefully, the builder is responsible for initialization of whole necessary fields when object is created.

Patch-set: 3