Update patch set 3
Patch Set 3:

Correction.  I do see a few superficial unit tests for the Cloudify version of this.  Bear in mind that this is a stretch goal and time is of the essence.  Let's create a JIRA for unit test coverage.  The other code convention comments are fine, but given their lack of application elsewhere, non-critical in this context.

Patch-set: 3