Update patch set 1
Patch Set 1:

Oddly.  It's a test that expects a ClassFormatException for some reason, and when it doesn't get it, it fails.  So maybe the question is "why would you have a unit test that expects a ClassFormatException".  Put another way, maybe things are *not* broken now, and that's what breaks the test.

Patch-set: 1