Remove wrapper from WorkflowResponse
When the workflow response succeeds, the requestReferences object is being wrapped by "WorkflowResponse": {"response": {... and all the other fields from the WorkflowResponse object are being included as well.  Adding a WorkflowResponseSerializer didn't seem to work since the root node was still included and there was no ObjectMapper in which to set WrapRootValue to false.  So this is the next best thing.

Patch 2 fixes most of the broken unit tests.

Change-Id: Ifa5bd02e70b23f41c9042ac207848c8ade77313a
Issue-ID: SO-586
Signed-off-by: Arthur Martella <am153x@att.com>