Update patch set 2
Patch Set 2:

I like the fix , but a little issue is the timing of the fix. This fix needs to be tested properly as this is the core of the functionality now and any disruption it might bring in at this moment would be very hazardous for ONAP :). Please ensure you have tested it and confirm.

Patch-set: 2