Update patch set 2
Patch Set 2:

Hi Seshu, I understand your concerns. I know it touches the key part of the code, however my changes are related to:
1. create wrapper for parameters to rest client - RequestClientParamater. It wraps parameters in CamundaClient into one object. Now, we pass one wrapper object instead of many parameters. It is cleaner.
2. Add to this wrapper one parameter more - correlationId, so we are able now to pass it to camunda process.

As you can see, the changes are not complicated and all covered with junit tests, I checked it. For me, junit for this kind of changes are enough. Of course we want to test it all in practice, in open lab. We are working on it.

Patch-set: 2