Update patch set 2
Patch Set 2:

I had to fix this change. I cannot merge with a parent as an abandoned patch. Please do not submit patches with parents that are not merged. 

Please read the wiki on how to submit a change and how to rebase that change on master before submitting and how to push a patch update to a change.

Patch-set: 2