Update patch set 1
Patch Set 1:

> (1 comment)

Hi Rob, I did it this way because in the previuous implementation (reading from properties file from resource project folder) when we want to change dmaap host and port, it causes we need to rebuild docker image, what is not good. 

What about reading from external file? That file would be place for example in etc folder and host and port for dmaap will be read from there each time when they are needed. 
Do you think is this better idea?

Patch-set: 1