Update patch set 1
Patch Set 1: -Code-Review

I believe there is already a way to read properties from an external file. The other parts of SO do this already, I don't remember where it is in the code. There are properties files for api-handler, adapters, and bpmn. The dmaap properties can be copied to one of those. If this work is not time sensitive I would defer it until containerized SO is released in a week or two.

Patch-set: 1
Reviewer: Gerrit User 97 <97@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0