Update patch set 1
Patch Set 1:

> (1 comment)

I analised the current properties implementation. It allows for configuration in mso-docker.json file, which is placed in docker-config repo. The property files are generated from this json file. The problem is that if we want to change configiuration in json file, we have to restart application. It would be good idea to have a possiblity to change config dynamically, without the need of restart. 
In the current code, we have WatchEvent implementation for changes property files for etc/mso/config.d. But it is not safe, because after restart these files are generated and all changes are lost. Putting dmaap host and port as enviromental variables let us have possibility for dynamic changes. Rob, what do you think?

Patch-set: 1