Update patch set 1
Patch Set 1:

> Why do we need dynamic configuration of a host name/port?  I am not
 > sure what the issue is with just re installing the docker image to
 > pick up the latest config is?

It would be more comfortable way to pick up the latest dmaap config without restart, what if we want to change for example dmaap ip and currently we can not restart so because some tasks are performed. Anyway, for now I can do it as it is, to use mso-docker.json file. If there is a need, I will come back to this topic in the future.

Patch-set: 1
Reviewer: Gerrit User 2789 <2789@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0