Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > I see no reason for this change - there should be no two same logs
 > in code anyway - otherwise that would mean that two parts of code
 > do same thing, and should be extracted to separate function
 > (actually that's the case here, code was clearly copied here)

Hi Lukasz,

In general when programming in Java, it is preferred to repeat a constant instead of repeating a literal. Although the program would run the same with or without it, this minor change allows the program to follow Java coding conventions. 

 > (1 comment)
 > 
 > I see no reason for this change - there should be no two same logs
 > in code anyway - otherwise that would mean that two parts of code
 > do same thing, and should be extracted to separate function
 > (actually that's the case here, code was clearly copied here)

Hi Lukasz,

In general, it is preferred to have a repeated constant instead of a repeated literal when programming in Java. All this change does is assigns a String variable to a phrase that is repeated throughout the program so a constant is repeated instead of a literal. Although the program would work equally well with or without it, this minor change allows the program to follow Java code conventions. 

Please let me know if you have any more questions or concerns.

Thanks,
Anisha

Patch-set: 1
Reviewer: Gerrit User 4423 <4423@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0