Update patch set 1
Patch Set 1:

We are refactoring the groovy code to remove the usage of AAI utils and instead use the AAI Client.  This allow us to have a central place to manage AAI changes.  The current issue we have is when AAI cuts a version we must manually modify many groovy files now to be compliant.  With this change all versioning is centralized to the AAI client.

Patch-set: 1