Groovy Compile
Groovy was not compiling as the maven plugin was off.  I had to move
some files around, and add some import statements to make it compile
again.  I also had to rename some packages to onap.so. A few files were
deleted that did not compile but were not in use.

Issue-ID: SO-823

Change-Id: I99d04b4f25709c86d397d72700c237d81362d8b3
Signed-off-by: Smokowski, Steve (ss835w) <ss835w@us.att.com>