Update patch set 2
Patch Set 2: Code-Review+2

Code review is one of the biggest strengths of Open Source, but it cannot be done if we don't conform to some basic ground rules. It's hard to appropriately review a patch that doesn't contain adequate detail in the commit message. It's hard to appropriately review a patch that does not conform it's subject to a single coherent area. It's hard to appropriately review a patch that is very large.

I know in ONAP these types of patches have been pushed through regardless in the past. I can't imagine that appropriate code review happened in the community on those patches. I'm hoping that we will collectively learn from longstanding and very successful open source projects and curtail patches that are not easy to review, or that are not review-able in the future. 

That said. I've reviewed this with the information provided by Steve. It seems fine.

Patch-set: 2
Reviewer: Gerrit User 667 <667@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2