Changed RequestFactory for Catalog DB client
- Missed some files, the repository isn't required now that we
dont fetch the identity service directly
- Fixing review comments and removing the unused import
- Fixing the build :) Now that cloudSite has a 1-1 relationship with
cloudidentity, there shouldn't be a need to query the identity on its
own unless the site is missing Refactored some code to use site to
fetch identity as a composition
- Changing the catalogDb client to use
HttpComponentsClientHttpRequestFactory instead of
SimpleClientHttpRequestFactory, the later would lead to a IOException
(FileNotFoundException specifically) whenever a 40X type of a response
code is returned from the spring rest endpoint. Adding some tests for
the catalogDbClient which earlier got missed in handover. Also fixing
another issue with cloudSite repository which was not fetching data
correctly. Fixed a repository which was added initially but later the
same went missing (Not sure how)

Change-Id: I72506865c4c5f7fd07a6031c146a4aba1a96c2c8
Issue-ID: SO-892
Signed-off-by: Benjamin, Max (mb388a) <mb388a@us.att.com>