Update patch set 1
Patch Set 1:

The jira issue describes the Nexus IQ findings.  I believe upgrading to 2.8.10 will solve this CVE:

CVE-2017-7525

A deserialization flaw was discovered in the jackson-databind, versions before 2.6.7.1, 2.7.9.1 and 2.8.9, which could allow an unauthenticated user to perform code execution by sending the maliciously crafted input to the readValue method of the ObjectMapper.
Explanation

Patch-set: 1