Update patch set 1
Patch Set 1: Code-Review-1

This is probably not a good change.  If there is actually a case where vpnRequestInputs could legitimately be null, and it makes logic sense to skip those lines of code, then it is ok.  If so, then please explain that in the JIRA ticket or in a code comment.  Otherwise it is just wrong.  An else case to handle the null would make it right.

Patch-set: 1
Reviewer: Gerrit User 1423 <1423@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1