Update patch set 5
Patch Set 5:

You can certainly override the default naming of the class name as the bean name, but in all the other bpmn flows we use the class name as is. For example, I have a Component class called "MyClass". If I wanted to use it in a bpmn flow as a bean I would use ${MyClass.myMethod()}.

Patch-set: 5
Label: Code-Review=0