Update patch set 6
Patch Set 6: Code-Review-2

I think this belongs in the casablanca branch without the .gitignore change.  We will merge casablanca commits to master as soon as we can work out the permissions with LF.  Also please be aware that we have an override.yaml file for so-monitoring in the so/docker-config repo as well as in the oom repo.  These will need to be updated too.

Patch-set: 6
Reviewer: Gerrit User 1423 <1423@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2