Make homingBB optinal and fix homing process
Properly defined whether or not to call homing at AssignVnfBB where by
default homing doesn't get called, but if a homing solution is provided,
then we resolve the solutions.
The goal is to change the BPMN definition to use serviceTask rather
than a scriptTask, allowing to use the Srping component to inject
logic.
Also we create a new HomingV2 component that is agnostic of the homing
implementation, which goal is to route the homing request to the proper
bbackend implementation; e.g. oof vs homing.
Finally, we remove the un-needed variables set in the execution context.

Change-Id: Ic9fb2a599c6375617c92a6d3462b62921ee47a6f
Issue-ID: SO-1168
Signed-off-by: Alexis de Talhouët <adetalhouet89@gmail.com>