Fix issues with mulicloud adapter parameters
Use the static final string variable for the parameter
key values instead of literal string of the variable name.

Also, vf_module_id and vnf_id (generic vnf id for multicloud)
are already part of the required input list, so they are
not needed to be added to the list of inputs.

Change-Id: I2b0bbf385e4114f0432673a3d7e8c9d652226f0a
Issue-ID: SO-1210
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>