Update patch set 3
Patch Set 3:

I appreciate the unit tests, but most of these old groovy utils will not continue to be in use.  I am not sure its worth the effort to add additional tests to them.  This flow in particular should not be used inside newer bpmn flows.

Patch-set: 3
Label: Code-Review=0