Update patch set 4
Patch Set 4:

Hi Steve,
Thank you for the information. So will this commit be merged or should we drop it completely?
Besides, don't you think that unused code should be thrown away or at least marked as deprecated? It would help not to waste time in the future and spend it on something more productive. This whole catalogue with groovy files seems to be a big mess with no tests, so hard to refactor.

Patch-set: 4
Label: Code-Review=0