Update patch set 5
Patch Set 5: -Code-Review

> Hi Steve,
 > Thank you for the information. So will this commit be merged or
 > should we drop it completely?
 > Besides, don't you think that unused code should be thrown away or
 > at least marked as deprecated? It would help not to waste time in
 > the future and spend it on something more productive. This whole
 > catalogue with groovy files seems to be a big mess with no tests,
 > so hard to refactor.

The future of this groovy files should be discussed. If they are deprecated, it is true that it is better to spend time on other, more productive tasks. Meanwhile, the ConfirmVolumeGroupName.groovy is actually used in bpmn flow and if tests are already done, I see no reason why they should not be merged.

Patch-set: 5