Update patch set 2
Patch Set 2: Code-Review-1

this does not fix the problem in SO-1358.
I see the '"' getting escaped in the catalog DB if I look directly with sql,
but what gets returned from the query has not escape and the json parsing still fails.

I will post more details in the JIRA

Patch-set: 2
Label: Code-Review=-1