Update patch set 7
Patch Set 7: Code-Review-1

Please convert the A&AI calls to use the SO standard A&AI client. I can try to find time to work with you on this. The SO A&AI client supports the single-transaction API in A&AI which allows you to group together any number of operations to execute as an entire set. A&AI will then manage the rollback for you in the case of a failure. 

An example can be found in the unit test here https://gerrit.onap.org/r/gitweb?p=so.git;a=blob;f=common/src/test/java/org/onap/so/client/aai/AAISingleTransactionClientTest.java;hb=b80f29cbbeb5ec218b0278deef872f7b5350e292

Patch-set: 7