Update patch set 4
Patch Set 4:

(2 comments)

If I recall correctly some of the interface details were not being brought in by the python script.

Munir - how similar or different is the AAI update functionality you've coded here vs. what is present in the python version of heatbridge?  That is, have have added anything that is not done by the python?

Patch-set: 4