Update patch set 7
Patch Set 7: Code-Review-1

I'd like to see that we support doing this AAI update in a generic way.  This code (with previously made comments on AAI client, etc.) is probably more or less fine.
I've written up a high level proposal for a generic approach.  I'm interested to get opinions/assessments on how much or little work it would take to support the generic approach and can we modify this gerrit so that it would line up with the generic approach (improvements on the generic proposal are also welcome). 

see here for the proposal: https://wiki.onap.org/pages/viewpage.action?pageId=58228881

Patch-set: 7
Label: Code-Review=-1