Update patch set 2
Patch Set 2:

> Did you really need to introduce grpc here as a client? 

Yes, reason for this is gRPC is HTTP2 based, hence I can have a streaming based communication between SO and CDS. Also, no need to manipulate URI, everything is based on .proto messages defined in CDS: https://github.com/onap/ccsdk-apps/tree/master/components/model-catalog/proto-definition/proto

Point of gRPC is, SO will send one request to CDS through that stream, and CDS will respond back with many sub-responses to update SO of the current status of the request. We're doing this because we were told SO is the place to have status of request being updated.

We had a few design session call with Seshu, Gil, Yuriy, and I thought even you were a part of it. We could hold on merging that code and have a follow-up call to

Patch-set: 2