Add cloudOwner to VNF adapter call
(Not a complete patch - for review and feedback)
The BPMN layer has the CloudRegion which includes the CloudOwner
attribute, but this is not passed to the VNF adapter layer.
Calls to the Multicloud infrastructure workload need the
CloudOwner as well.

This patch sketches out what it might look like to add the CloudOwner
to the VNF adapter API.

Change-Id: I88028096fbd84fc3fe5d53786c70b3c8806bb711
Issue-ID: SO-1447
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>