Update patch set 3
Patch Set 3:

> Was was this change made?

Hi Lukasz,

instead of running the thread to do message correlation, I changed to use spring applicationEvent to notify about the incoming message. 

The main issue with runnable.run() is the thread is not released( =null), and all the following activities will be executed in both the main thread and the thread started by runnable.run().  

I haven't gotten the stable environment to test it e2e, and I am trying to set that up. 

there are few code change i need to make, I will raise a bug to track that.

Patch-set: 3