Update patch set 6
Patch Set 6: -Code-Review

(1 comment)

I have to admit I missed the discussion regarding updating orchestration flow table, along with the why it is breaking backward compatibility.

As I understand with this patch, we're manually augmenting the orchestration flow with our new BB instead of relying on the SQL driven flow.
It doesn't seem the best approach, but if this is the right approach, then we should track this as a tech dept to address later point in time, with proper design - as this looks like a hack.

Patch-set: 6
Reviewer: Gerrit User 137 <137@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0