Fix GR_API Homing
This change uses the correct
variable name to enable homing
in AssignVnfBB.
The correct variable to use is
not 'callHoming' but 'homing'.

This change also corrects HomingBB
and calls HomingV2 rather than SniroV2
or OofV2. HomingV2 already passes calls
to the correct implimentation, no need
do this in the BPMN.

Issue-ID: SO-1470
Change-Id: I14a77ef75b8b86feed7af3aab4d52c0df5f12886
Signed-off-by: Marcus G K Williams <marcus.williams@intel.com>