Update patch set 5
Patch Set 5: -Code-Review

In my laptop, it's failing straightaway and then wait for the timeout. If you debug the test, it's passing almost every time. I have one thought, it might depend on the time workflowMessage is inited in the BPMN execution. that should before the 'inform' call.

Patch-set: 5
Label: Code-Review=0